feat: 合约对象抽离
